requirements.txt
streamlit==1.43.2
langchain==0.3.21
langchain-community==0.3.20
faiss-cpu==1.10.0
sentence-transformers==3.4.1
pdfplumber==0.11.5
python-dotenv==1.1.0
